Refactor post-build.sh and remove backup files created by sed
The -i flag for sed will create a backup file which we do not need.
Specifying -i "" (empty quotes) leads to no backup files.
This has the nice side effect of lowering our bundle size since we ended
up with .js-e files which were the backup files of the sed replacements